REFACTOR :: header component 분ㄹㅣ(#82)
